+++ apps/emqttd/src/emqttd_broker.erl
@@ -39,8 +39,14 @@
 %% API Function Exports
 -export([start_link/0]).
 
+%% Running nodes
+-export([running_nodes/0]).
+
 %% Event API
 -export([subscribe/1, notify/2]).
+
+%% Hook API
+-export([hook/3, unhook/2, foreach_hooks/2, foldl_hooks/3]).
 
 %% Broker API
 -export([env/1, version/0, uptime/0, datetime/0, sysdescr/0]).
@@ -67,6 +73,13 @@
 -spec start_link() -> {ok, pid()} | ignore | {error, any()}.
 start_link() ->
     gen_server:start_link({local, ?SERVER}, ?MODULE, [], []).
+
+%%------------------------------------------------------------------------------
+%% @doc Get running nodes
+%% @end
+%%------------------------------------------------------------------------------
+running_nodes() ->
+    mnesia:system_info(running_db_nodes).
 
 %%------------------------------------------------------------------------------
 %% @doc Subscribe broker event
@@ -128,6 +141,52 @@ datetime() ->
             "~4..0w-~2..0w-~2..0w ~2..0w:~2..0w:~2..0w", [Y, M, D, H, MM, S])).
 
 %%------------------------------------------------------------------------------
+%% @doc Hook
+%% @end
+%%------------------------------------------------------------------------------
+-spec hook(Hook :: atom(), Name :: any(), MFA :: mfa()) -> ok | {error, any()}.
+hook(Hook, Name, MFA) ->
+    gen_server:call(?MODULE, {hook, Hook, Name, MFA}).
+
+%%------------------------------------------------------------------------------
+%% @doc Unhook
+%% @end
+%%------------------------------------------------------------------------------
+-spec unhook(Hook :: atom(), Name :: any()) -> ok | {error, any()}.
+unhook(Hook, Name) ->
+    gen_server:call(?MODULE, {unhook, Hook, Name}).
+
+%%------------------------------------------------------------------------------
+%% @doc Foreach hooks
+%% @end
+%%------------------------------------------------------------------------------
+-spec foreach_hooks(Hook :: atom(), Args :: list()) -> any().
+foreach_hooks(Hook, Args) ->
+    case ets:lookup(?BROKER_TAB, {hook, Hook}) of
+        [{_, Hooks}] ->
+            lists:foreach(fun({_Name, {M, F, A}}) ->
+                    apply(M, F, Args++A)
+                end, Hooks);
+        [] ->
+            ok
+    end.
+
+%%------------------------------------------------------------------------------
+%% @doc Foldl hooks
+%% @end
+%%------------------------------------------------------------------------------
+-spec foldl_hooks(Hook :: atom(), Args :: list(), Acc0 :: any()) -> any().
+foldl_hooks(Hook, Args, Acc0) ->
+    case ets:lookup(?BROKER_TAB, {hook, Hook}) of
+        [{_, Hooks}] -> 
+            lists:foldl(fun({_Name, {M, F, A}}, Acc) -> 
+                    apply(M, F, [Acc, Args++A])
+                end, Acc0, Hooks);
+        [] -> 
+            Acc0
+    end.
+
+%%------------------------------------------------------------------------------
 %% @doc Start a tick timer
 %% @end
 %%------------------------------------------------------------------------------
@@ -156,6 +215,7 @@ init([]) ->
     random:seed(now()),
     ets:new(?BROKER_TAB, [set, public, named_table]),
     % Create $SYS Topics
+    emqttd_pubsub:create(<<"$SYS/brokers">>),
     [ok = create_topic(Topic) || Topic <- ?SYSTOP_BROKERS],
     % Tick
     {ok, #state{started_at = os:timestamp(), tick_tref = start_tick(tick)}, hibernate}.
@@ -163,13 +223,39 @@ init([]) ->
 handle_call(uptime, _From, State) ->
     {reply, uptime(State), State};
 
+handle_call({hook, Hook, Name, MFArgs}, _From, State) ->
+    Key = {hook, Hook}, Reply =
+    case ets:lookup(?BROKER_TAB, Key) of
+        [{Key, Hooks}] -> 
+            case lists:keyfind(Name, 1, Hooks) of
+                {Name, _MFArgs} ->
+                    {error, existed};
+                false ->
+                    ets:insert(?BROKER_TAB, {Key, Hooks ++ [{Name, MFArgs}]})
+            end;
+        [] -> 
+            ets:insert(?BROKER_TAB, {Key, [{Name, MFArgs}]})
+    end,
+    {reply, Reply, State};
+
+handle_call({unhook, Name}, _From, State) ->
+    Key = {hook, Name}, Reply =
+    case ets:lookup(?BROKER_TAB, Key) of
+        [{Key, Hooks}] -> 
+            ets:insert(?BROKER_TAB, {Key, lists:keydelete(Name, 1, Hooks)}); 
+        [] -> 
+            {error, not_found}
+    end,
+    {reply, Reply, State};
+
 handle_call(_Request, _From, State) ->
-    {reply, error, State}.
+    {reply, {error, unsupport_request}, State}.
 
 handle_cast(_Msg, State) ->
     {noreply, State}.
 
 handle_info(tick, State) ->
+    retain(brokers),
     retain(version, list_to_binary(version())),
     retain(sysdescr, list_to_binary(sysdescr())),
     publish(uptime, list_to_binary(uptime(State))),
@@ -191,6 +277,10 @@ code_change(_OldVsn, State, _Extra) ->
 
 create_topic(Topic) ->
     emqttd_pubsub:create(emqtt_topic:systop(Topic)).
+
+retain(brokers) ->
+    Payload = list_to_binary(string:join([atom_to_list(N) || N <- running_nodes()], ",")),
+    publish(#mqtt_message{retain = true, topic = <<"$SYS/brokers">>, payload = Payload}).
 
 retain(Topic, Payload) when is_binary(Payload) ->
     publish(#mqtt_message{retain = true,