+++ apps/emqttd/src/emqttd_mod_autosub.erl
@@ -1,5 +1,5 @@
 %%%-----------------------------------------------------------------------------
-%%% @Copyright (C) 2012-2015, Feng Lee <feng@emqtt.io>
+%%% Copyright (c) 2012-2015 eMQTT.IO, All Rights Reserved.
 %%%
 %%% Permission is hereby granted, free of charge, to any person obtaining a copy
 %%% of this software and associated documentation files (the "Software"), to deal
@@ -20,23 +20,41 @@
 %%% SOFTWARE.
 %%%-----------------------------------------------------------------------------
 %%% @doc
-%%% presence manager application
+%%% emqttd auto subscribe module.
 %%%
 %%% @end
 %%%-----------------------------------------------------------------------------
--module(emqttd_presence_app).
 
--behaviour(application).
+-module(emqttd_mod_autosub).
 
-%% Application callbacks
--export([start/2, stop/1]).
+-author("Feng Lee <feng@emqtt.io>").
 
-%% ===================================================================
-%% Application callbacks
-%% ===================================================================
+-include_lib("emqtt/include/emqtt.hrl").
 
-start(_StartType, _StartArgs) ->
-    emqttd_presence_sup:start_link().
+-include_lib("emqtt/include/emqtt_packet.hrl").
 
-stop(_State) ->
-    ok.
+-include("emqttd.hrl").
+
+-behaviour(emqttd_gen_mod).
+
+-export([load/1, client_connected/3, unload/1]).
+
+-record(state, {topics}).
+
+load(Opts) ->
+    Topics = [{list_to_binary(Topic), Qos} || {Topic, Qos} <- Opts, 0 =< Qos, Qos =< 2],
+    emqttd_broker:hook(client_connected, {?MODULE, client_connected},
+                       {?MODULE, client_connected, [Topics]}),
+    {ok, #state{topics = Topics}}.
+
+client_connected(?CONNACK_ACCEPT, #mqtt_client{clientid = ClientId, client_pid = ClientPid}, Topics) ->
+    F = fun(Topic) -> emqtt_topic:feed_var(<<"$c">>, ClientId, Topic) end,
+    [ClientPid ! {subscribe, F(Topic), Qos} || {Topic, Qos} <- Topics];
+
+client_connected(_ConnAck, _Client, _Topics) ->
+    ignore.
+
+unload(_Opts) ->
+    emqttd_broker:unhook(client_connected, {?MODULE, client_connected}).
+
+