+++ apps/emqttd/src/emqttd_mod_sup.erl
@@ -1,5 +1,5 @@
 %%%-----------------------------------------------------------------------------
-%%% @Copyright (C) 2012-2015, Feng Lee <feng@emqtt.io>
+%%% Copyright (c) 2012-2015 eMQTT.IO, All Rights Reserved.
 %%%
 %%% Permission is hereby granted, free of charge, to any person obtaining a copy
 %%% of this software and associated documentation files (the "Software"), to deal
@@ -20,14 +20,48 @@
 %%% SOFTWARE.
 %%%-----------------------------------------------------------------------------
 %%% @doc
-%%% emqttd management dashboard.
+%%% emqttd module supervisor.
 %%%
 %%% @end
 %%%-----------------------------------------------------------------------------
--module(emqttd_dashboard).
+-module(emqttd_mod_sup).
 
 -author("Feng Lee <feng@emqtt.io>").
 
-%%TODO...
+-include("emqttd.hrl").
 
+-behaviour(supervisor).
+
+%% API
+-export([start_link/0, start_child/1, start_child/2]).
+
+%% Supervisor callbacks
+-export([init/1]).
+
+%% Helper macro for declaring children of supervisor
+-define(CHILD(Mod, Type), {Mod, {Mod, start_link, []}, permanent, 5000, Type, [Mod]}).
+
+%%%=============================================================================
+%%% API
+%%%=============================================================================
+
+start_link() ->
+    supervisor:start_link({local, ?MODULE}, ?MODULE, []).
+
+start_child(ChildSpec) when is_tuple(ChildSpec) ->
+	supervisor:start_child(?MODULE, ChildSpec).
+
+%%
+%% start_child(Mod::atom(), Type::type()) -> {ok, pid()}
+%% @type type() = worker | supervisor
+%%
+start_child(Mod, Type) when is_atom(Mod) and is_atom(Type) ->
+	supervisor:start_child(?MODULE, ?CHILD(Mod, Type)).
+
+%%%=============================================================================
+%%% Supervisor callbacks
+%%%=============================================================================
+
+init([]) ->
+    {ok, {{one_for_one, 10, 3600}, []}}.