+++ apps/emqttd/src/emqttd_protocol.erl
@@ -29,16 +29,17 @@
 -author("Feng Lee <feng@emqtt.io>").
 
 -include_lib("emqtt/include/emqtt.hrl").
+
 -include_lib("emqtt/include/emqtt_packet.hrl").
 
 -include("emqttd.hrl").
 
 %% API
--export([init/3, clientid/1]).
+-export([init/3, info/1, clientid/1, client/1]).
 
 -export([received/2, send/2, redeliver/2, shutdown/2]).
 
--export([info/1]).
+-export([handle/2]).
 
 %% Protocol State
 -record(proto_state, {
@@ -47,30 +48,29 @@
         connected = false, %received CONNECT action?
         proto_ver,
         proto_name,
-		%packet_id,
         username,
 		clientid,
 		clean_sess,
-        session, %% session state or session pid
+        session,    %% session state or session pid
 		will_msg,
-        max_clientid_len = ?MAX_CLIENTID_LEN
+        max_clientid_len = ?MAX_CLIENTID_LEN,
+        client_pid
 }).
 
 -type proto_state() :: #proto_state{}.
 
+%%------------------------------------------------------------------------------
+%% @doc Init protocol
+%% @end
+%%------------------------------------------------------------------------------
 init(Peername, SendFun, Opts) ->
     MaxLen = proplists:get_value(max_clientid_len, Opts, ?MAX_CLIENTID_LEN),
 	#proto_state{
         peername         = Peername,
         sendfun          = SendFun,
-        max_clientid_len = MaxLen}. 
+        max_clientid_len = MaxLen,
+        client_pid       = self()}. 
 
-clientid(#proto_state{clientid = ClientId}) -> ClientId.
-
-client(#proto_state{peername = {Addr, _Port}, clientid = ClientId, username = Username}) ->
-    #mqtt_client{clientid = ClientId, username = Username, ipaddr = Addr}.
-
-%%SHOULD be registered in emqttd_cm
 info(#proto_state{proto_ver    = ProtoVer,
                   proto_name   = ProtoName,
 				  clientid	   = ClientId,
@@ -78,11 +78,27 @@ info(#proto_state{proto_ver    = ProtoVer,
 				  will_msg	   = WillMsg}) ->
 	[{proto_ver,  ProtoVer},
      {proto_name, ProtoName},
-	 {clientid,  ClientId},
+	 {clientid,   ClientId},
 	 {clean_sess, CleanSess},
 	 {will_msg,   WillMsg}].
 
-%%CONNECT – Client requests a connection to a Server
+clientid(#proto_state{clientid = ClientId}) ->
+    ClientId.
+
+client(#proto_state{peername = {Addr, _Port},
+                    clientid = ClientId,
+                    username = Username,
+                    clean_sess = CleanSess,
+                    proto_ver  = ProtoVer,
+                    client_pid = Pid}) ->
+    #mqtt_client{clientid   = ClientId,
+                 username   = Username,
+                 ipaddress  = Addr,
+                 clean_sess = CleanSess,
+                 proto_ver  = ProtoVer,
+                 client_pid = Pid}.
+
+%% CONNECT – Client requests a connection to a Server
 
 %%A Client can only send the CONNECT Packet once over a Network Connection. 
 -spec received(mqtt_packet(), proto_state()) -> {ok, proto_state()} | {error, any()}. 
@@ -105,51 +121,62 @@ received(Packet = ?PACKET(_Type), State) ->
 		{error, Reason, State}
 	end.
 
-handle(Packet = ?CONNECT_PACKET(Var), State = #proto_state{peername = Peername = {Addr, _}}) ->
+handle(Packet = ?CONNECT_PACKET(Var), State0 = #proto_state{peername = Peername}) ->
 
     #mqtt_packet_connect{proto_ver  = ProtoVer,
+                         proto_name = ProtoName, 
                          username   = Username,
                          password   = Password,
                          clean_sess = CleanSess,
                          keep_alive = KeepAlive,
-                         clientid  = ClientId} = Var,
+                         clientid   = ClientId} = Var,
 
-    trace(recv, Packet, State#proto_state{clientid = ClientId}), %%TODO: fix later...
+    State1 = State0#proto_state{proto_ver  = ProtoVer,
+                                proto_name = ProtoName, 
+                                username   = Username,
+                                clientid   = ClientId,
+                                clean_sess = CleanSess},
 
-    State1 = State#proto_state{proto_ver  = ProtoVer,
-                               username   = Username,
-                               clientid  = ClientId,
-                               clean_sess = CleanSess},
-    {ReturnCode1, State2} =
-    case validate_connect(Var, State) of
+    trace(recv, Packet, State1),
+
+    {ReturnCode1, State3} =
+    case validate_connect(Var, State1) of
         ?CONNACK_ACCEPT ->
-            Client = #mqtt_client{clientid = ClientId, username = Username, ipaddr = Addr},
-            case emqttd_access_control:auth(Client, Password) of
+            case emqttd_access_control:auth(client(State1), Password) of
                 ok ->
-                    ClientId1 = clientid(ClientId, State),
+                    %% Generate clientId if null
+                    State2 = State1#proto_state{clientid = clientid(ClientId, State1)},
+
+                    %% Register the client to cm
+                    emqttd_cm:register(client(State2)),
+
+                    %%Starting session
+                    {ok, Session} = emqttd_session:start({CleanSess, clientid(State2), self()}),
+
+                    %% Start keepalive
                     start_keepalive(KeepAlive),
-                    emqttd_cm:register(ClientId1),
-                    {?CONNACK_ACCEPT, State1#proto_state{clientid  = ClientId1,
-                                                         will_msg   = willmsg(Var)}};
+
+                    %% ACCEPT
+                    {?CONNACK_ACCEPT, State2#proto_state{session = Session, will_msg = willmsg(Var)}};
                 {error, Reason}->
-                    lager:error("~s@~s: username '~s' login failed - ~s", [ClientId, emqttd_net:format(Peername), Username, Reason]),
+                    lager:error("~s@~s: username '~s', login failed - ~s",
+                                    [ClientId, emqttd_net:format(Peername), Username, Reason]),
                     {?CONNACK_CREDENTIALS, State1}
                                                         
             end;
         ReturnCode ->
             {ReturnCode, State1}
     end,
-    notify(connected, ReturnCode1, State2),
-    send(?CONNACK_PACKET(ReturnCode1), State2),
-    %%Starting session
-    {ok, Session} = emqttd_session:start({CleanSess, ClientId, self()}),
-    {ok, State2#proto_state{session = Session}};
+    %% Run hooks
+    emqttd_broker:foreach_hooks(client_connected, [ReturnCode1, client(State3)]),
+    %% Send connack
+    send(?CONNACK_PACKET(ReturnCode1), State3);
 
 handle(Packet = ?PUBLISH_PACKET(?QOS_0, Topic, _PacketId, _Payload),
        State = #proto_state{clientid = ClientId, session = Session}) ->
     case check_acl(publish, Topic, State) of
         allow -> 
-            emqttd_session:publish(Session, ClientId, {?QOS_0, emqtt_message:from_packet(Packet)});
+            do_publish(Session, ClientId, ?QOS_0, Packet);
         deny -> 
             lager:error("ACL Deny: ~s cannot publish to ~s", [ClientId, Topic])
     end,
@@ -159,7 +186,7 @@ handle(Packet = ?PUBLISH_PACKET(?QOS_1, Topic, PacketId, _Payload),
          State = #proto_state{clientid = ClientId, session = Session}) ->
     case check_acl(publish, Topic, State) of
         allow -> 
-            emqttd_session:publish(Session, ClientId, {?QOS_1, emqtt_message:from_packet(Packet)}),
+            do_publish(Session, ClientId, ?QOS_1, Packet),
             send(?PUBACK_PACKET(?PUBACK, PacketId), State);
         deny -> 
             lager:error("ACL Deny: ~s cannot publish to ~s", [ClientId, Topic]),
@@ -170,7 +197,7 @@ handle(Packet = ?PUBLISH_PACKET(?QOS_2, Topic, PacketId, _Payload),
          State = #proto_state{clientid = ClientId, session = Session}) ->
     case check_acl(publish, Topic, State) of
         allow -> 
-            NewSession = emqttd_session:publish(Session, ClientId, {?QOS_2, emqtt_message:from_packet(Packet)}),
+            NewSession = do_publish(Session, ClientId, ?QOS_2, Packet), 
             send(?PUBACK_PACKET(?PUBREC, PacketId), State#proto_state{session = NewSession});
         deny -> 
             lager:error("ACL Deny: ~s cannot publish to ~s", [ClientId, Topic]),
@@ -203,17 +230,23 @@ handle(?SUBSCRIBE_PACKET(PacketId, TopicTable), State = #proto_state{clientid = 
             lager:error("SUBSCRIBE from '~s' Denied: ~p", [ClientId, TopicTable]),
             {ok, State};
         false ->
+            TopicTable1 = emqttd_broker:foldl_hooks(client_subscribe, [], TopicTable),
             %%TODO: GrantedQos should be renamed.
-            {ok, NewSession, GrantedQos} = emqttd_session:subscribe(Session, TopicTable),
+            {ok, NewSession, GrantedQos} = emqttd_session:subscribe(Session, TopicTable1),
             send(?SUBACK_PACKET(PacketId, GrantedQos), State#proto_state{session = NewSession})
     end;
+
+handle({subscribe, Topic, Qos}, State = #proto_state{session = Session}) ->
+    {ok, NewSession, _GrantedQos} = emqttd_session:subscribe(Session, [{Topic, Qos}]),
+    {ok, State#proto_state{session = NewSession}};
 
 %% protect from empty topic list
 handle(?UNSUBSCRIBE_PACKET(PacketId, []), State) ->
     send(?UNSUBACK_PACKET(PacketId), State);
 
 handle(?UNSUBSCRIBE_PACKET(PacketId, Topics), State = #proto_state{session = Session}) ->
-    {ok, NewSession} = emqttd_session:unsubscribe(Session, Topics),
+    Topics1 = emqttd_broker:foldl_hooks(client_unsubscribe, [], Topics),
+    {ok, NewSession} = emqttd_session:unsubscribe(Session, Topics1),
     send(?UNSUBACK_PACKET(PacketId), State#proto_state{session = NewSession});
 
 handle(?PACKET(?PINGREQ), State) ->
@@ -224,6 +257,10 @@ handle(?PACKET(?DISCONNECT), State) ->
     % clean willmsg
     {stop, normal, State#proto_state{will_msg = undefined}}.
 
+do_publish(Session, ClientId, Qos, Packet) ->
+    Message = emqttd_broker:foldl_hooks(client_publish, [], emqtt_message:from_packet(Packet)),
+    emqttd_session:publish(Session, ClientId, {Qos, Message}).
+
 -spec send({pid() | tuple(), mqtt_message()} | mqtt_packet(), proto_state()) -> {ok, proto_state()}.
 %% qos0 message
 send({_From, Message = #mqtt_message{qos = ?QOS_0}}, State) ->
@@ -232,7 +269,6 @@ send({_From, Message = #mqtt_message{qos = ?QOS_0}}, State) ->
 %% message from session
 send({_From = SessPid, Message}, State = #proto_state{session = SessPid}) when is_pid(SessPid) ->
 	send(emqtt_message:to_packet(Message), State);
-
 %% message(qos1, qos2) not from session
 send({_From, Message = #mqtt_message{qos = Qos}}, State = #proto_state{session = Session}) 
     when (Qos =:= ?QOS_1) orelse (Qos =:= ?QOS_2) ->
@@ -260,28 +296,40 @@ trace(send, Packet, #proto_state{peername  = Peername, clientid = ClientId}) ->
 redeliver({?PUBREL, PacketId}, State) ->
     send(?PUBREL_PACKET(PacketId), State).
 
+shutdown(duplicate_id, _State) ->
+    quiet; %%
+
+shutdown(_, #proto_state{clientid = undefined}) ->
+    ignore;
+
 shutdown(Error, #proto_state{peername = Peername, clientid = ClientId, will_msg = WillMsg}) ->
-    send_willmsg(ClientId, WillMsg),
-    try_unregister(ClientId, self()),
-	lager:info([{client, ClientId}], "Protocol ~s@~s Shutdown: ~p",
+	lager:info([{client, ClientId}], "Client ~s@~s: shutdown ~p",
                    [ClientId, emqttd_net:format(Peername), Error]),
-    ok.
+    send_willmsg(ClientId, WillMsg),
+    try_unregister(ClientId),
+    emqttd_broker:foreach_hooks(client_disconnected, [Error, ClientId]).
 
 willmsg(Packet) when is_record(Packet, mqtt_packet_connect) ->
     emqtt_message:from_packet(Packet).
 
+%% generate a clientId
+clientid(undefined, State) ->
+    clientid(<<>>, State);
+%%TODO: <<>> is not right.
 clientid(<<>>, #proto_state{peername = Peername}) ->
-    <<"eMQTT_", (base64:encode(emqttd_net:format(Peername)))/binary>>;
-
+    {_, _, MicroSecs} = os:timestamp(),
+    iolist_to_binary(["emqttd_", base64:encode(emqttd_net:format(Peername)), integer_to_list(MicroSecs)]);
 clientid(ClientId, _State) -> ClientId.
 
 send_willmsg(_ClientId, undefined) ->
     ignore;
 %%TODO:should call session...
 send_willmsg(ClientId, WillMsg) -> 
+    lager:info("Client ~s send willmsg: ~p", [ClientId, WillMsg]),
     emqttd_pubsub:publish(ClientId, WillMsg).
 
 start_keepalive(0) -> ignore;
+
 start_keepalive(Sec) when Sec > 0 ->
     self() ! {keepalive, start, round(Sec * 1.5)}.
 
@@ -353,8 +401,8 @@ validate_qos(undefined) -> true;
 validate_qos(Qos) when Qos =< ?QOS_2 -> true;
 validate_qos(_) -> false.
 
-try_unregister(undefined, _) -> ok;
-try_unregister(ClientId, _) -> emqttd_cm:unregister(ClientId).
+try_unregister(undefined) -> ok;
+try_unregister(ClientId)  -> emqttd_cm:unregister(ClientId).
 
 %% publish ACL is cached in process dictionary.
 check_acl(publish, Topic, State) ->
@@ -386,19 +434,4 @@ inc(?PINGRESP) ->
     emqttd_metrics:inc('packets/pingresp');
 inc(_) ->
     ingore.
-
-notify(connected, ReturnCode, #proto_state{peername   = Peername, 
-                                           proto_ver  = ProtoVer, 
-                                           clientid  = ClientId, 
-                                           clean_sess = CleanSess}) ->
-    Sess = case CleanSess of
-        true -> false;
-        false -> true
-    end,
-    Params = [{from, emqttd_net:format(Peername)},
-              {protocol, ProtoVer},
-              {session, Sess},
-              {connack, ReturnCode}],
-    emqttd_event:notify({connected, ClientId, Params}).
-