+++ apps/emqttd/src/emqttd_pubsub.erl
@@ -24,6 +24,7 @@
 %%%
 %%% @end
 %%%-----------------------------------------------------------------------------
+
 -module(emqttd_pubsub).
 
 -author("Feng Lee <feng@emqtt.io>").
@@ -64,6 +65,12 @@
 %%% Mnesia callbacks
 %%%=============================================================================
 mnesia(boot) ->
+    %% p2p queue table
+    ok = emqttd_mnesia:create_table(queue, [
+                {type, set},
+                {ram_copies, [node()]},
+                {record_name, mqtt_queue},
+                {attributes, record_info(fields, mqtt_queue)}]),
     %% topic table
     ok = emqttd_mnesia:create_table(topic, [
                 {type, bag},
@@ -80,6 +87,7 @@ mnesia(boot) ->
                 {local_content, true}]);
 
 mnesia(copy) ->
+    ok = emqttd_mnesia:copy_table(queue),
     ok = emqttd_mnesia:copy_table(topic),
     ok = emqttd_mnesia:copy_table(subscriber).
 
@@ -102,6 +110,9 @@ start_link(Id, Opts) ->
 %% @end
 %%------------------------------------------------------------------------------
 -spec create(Topic :: binary()) -> ok | {error, Error :: any()}.
+create(<<"$Q/", _Queue/binary>>) ->
+    %% protecte from queue
+    {error, cannot_create_queue};
 create(Topic) when is_binary(Topic) ->
     TopicR = #mqtt_topic{topic = Topic, node = node()},
     case mnesia:transaction(fun add_topic/1, [TopicR]) of
@@ -148,8 +159,8 @@ cast(Msg) ->
 %% @end
 %%------------------------------------------------------------------------------
 -spec publish(From :: mqtt_clientid() | atom(), Msg :: mqtt_message()) -> ok.
-publish(From, Msg=#mqtt_message{topic=Topic}) ->
-    lager:info([{client, From}, {topic, Topic}], "~s PUBLISH to ~s", [From, Topic]),
+publish(From, #mqtt_message{topic=Topic} = Msg) ->
+    trace(publish, From, Msg),
     %% Retain message first. Don't create retained topic.
     case emqttd_msg_store:retain(Msg) of
         ok ->
@@ -158,6 +169,16 @@ publish(From, Msg=#mqtt_message{topic=Topic}) ->
         ignore ->
             publish(From, Topic, Msg)
      end.
+
+publish(From, <<"$Q/", _/binary>> = Queue, #mqtt_message{qos = Qos} = Msg) ->
+    lists:foreach(
+        fun(#mqtt_queue{subpid = SubPid, qos = SubQos}) -> 
+            Msg1 = if
+                Qos > SubQos -> Msg#mqtt_message{qos = SubQos};
+                true -> Msg
+            end,
+            SubPid ! {dispatch, {self(), Msg1}}
+        end, mnesia:dirty_read(queue, Queue));
     
 publish(_From, Topic, Msg) when is_binary(Topic) ->
 	lists:foreach(fun(#mqtt_topic{topic=Name, node=Node}) ->
@@ -172,7 +193,7 @@ publish(_From, Topic, Msg) when is_binary(Topic) ->
 %% @end
 %%------------------------------------------------------------------------------
 -spec dispatch(Topic :: binary(), Msg :: mqtt_message()) -> non_neg_integer().
-dispatch(Topic, Msg = #mqtt_message{qos = Qos}) when is_binary(Topic) ->
+dispatch(Topic, #mqtt_message{qos = Qos} = Msg ) when is_binary(Topic) ->
     Subscribers = mnesia:dirty_read(subscriber, Topic),
     setstats(dropped, Subscribers =:= []), %%TODO:...
     lists:foreach(
@@ -197,23 +218,44 @@ match(Topic) when is_binary(Topic) ->
 init([Id, _Opts]) ->
     process_flag(min_heap_size, 1024*1024),
     gproc_pool:connect_worker(pubsub, {?MODULE, Id}),
+    %%TODO: gb_trees to replace maps?
     {ok, #state{id = Id, submap = maps:new()}}.
 
 handle_call({subscribe, SubPid, Topics}, _From, State) ->
-    TopicSubs = lists:map(fun({Topic, Qos}) ->
-                    {#mqtt_topic{topic = Topic, node = node()},
-                     #mqtt_subscriber{topic = Topic, qos = Qos, pid = SubPid}}
+    TopicSubs = lists:map(fun({<<"$Q/", _/binary>> = Queue, Qos}) ->
+                            #mqtt_queue{name = Queue, subpid = SubPid, qos = Qos};
+                             ({Topic, Qos}) ->
+                            {#mqtt_topic{topic = Topic, node = node()},
+                             #mqtt_subscriber{topic = Topic, qos = Qos, pid = SubPid}}
                 end, Topics),
     F = fun() ->
-            lists:map(fun add_subscriber/1, TopicSubs) 
+            lists:map(fun(QueueR) when is_record(QueueR, mqtt_queue) ->
+                        add_queue(QueueR);
+                         (TopicSub) ->
+                        add_subscriber(TopicSub)
+                end, TopicSubs)
         end,
     case mnesia:transaction(F) of
         {atomic, _Result} ->
             setstats(all),
             NewState = monitor_subscriber(SubPid, State),
-            %% grant all qos
+            %%TODO: grant all qos
             {reply, {ok, [Qos || {_Topic, Qos} <- Topics]}, NewState};
         {aborted, Error} ->
+            {reply, {error, Error}, State}
+    end;
+
+handle_call({subscribe, SubPid, <<"$Q/", _/binary>> = Queue, Qos}, _From, State) ->
+    case mnesia:dirty_read(queue, Queue) of
+        [OldQueueR] -> lager:error("Queue is overwrited by ~p: ~p", [SubPid, OldQueueR]);
+        [] -> ok
+    end,
+    QueueR = #mqtt_queue{name = Queue, subpid = SubPid, qos = Qos},
+    case mnesia:transaction(fun add_queue/1, [QueueR]) of
+        {atomic, ok} ->
+            setstats(queues),
+            {reply, {ok, Qos}, monitor_subscriber(SubPid, State)};
+        {aborted, Error} -> 
             {reply, {error, Error}, State}
     end;
 
@@ -233,16 +275,36 @@ handle_call(Req, _From, State) ->
 	{reply, {error, badreq}, State}.
 
 handle_cast({unsubscribe, SubPid, Topics}, State) when is_list(Topics) ->
-    TopicSubs = lists:map(fun(Topic) ->
-                    {#mqtt_topic{topic = Topic, node = node()},
-                     #mqtt_subscriber{topic = Topic, _ = '_', pid = SubPid}}
+
+    TopicSubs = lists:map(fun(<<"$Q/", _/binary>> = Queue) ->
+                                #mqtt_queue{name = Queue, subpid = SubPid};
+                             (Topic) ->
+                                {#mqtt_topic{topic = Topic, node = node()},
+                                 #mqtt_subscriber{topic = Topic, _ = '_', pid = SubPid}}
                 end, Topics),
-    F = fun() -> lists:foreach(fun remove_subscriber/1, TopicSubs) end,
+    F = fun() ->
+            lists:foreach(
+                fun(QueueR) when is_record(QueueR, mqtt_queue) ->
+                        remove_queue(QueueR);
+                    (TopicSub) ->
+                        remove_subscriber(TopicSub)
+                end, TopicSubs) 
+        end,
     case mnesia:transaction(F) of
         {atomic, _} -> ok;
         {aborted, Error} -> lager:error("unsubscribe ~p error: ~p", [Topics, Error])
     end,
     setstats(all),
+    {noreply, State};
+
+handle_cast({unsubscribe, SubPid, <<"$Q/", _/binary>> = Queue}, State) ->
+    QueueR = #mqtt_queue{name = Queue, subpid = SubPid},
+    case mnesia:transaction(fun remove_queue/1, [QueueR]) of
+        {atomic, _} -> 
+            setstats(queues);
+        {aborted, Error} -> 
+            lager:error("unsubscribe queue ~s error: ~p", [Queue, Error])
+    end,
     {noreply, State};
 
 handle_cast({unsubscribe, SubPid, Topic}, State) ->
@@ -264,6 +326,13 @@ handle_info({'DOWN', _Mon, _Type, DownPid, _Info}, State = #state{submap = SubMa
         true ->
             Node = node(),
             F = fun() -> 
+                    %% remove queue...
+                    Queues = mnesia:match_object(queue, #mqtt_queue{subpid = DownPid, _ = '_'}, write),
+                    lists:foreach(fun(QueueR) ->
+                                mnesia:delete_object(queue, QueueR, write) 
+                        end, Queues),
+
+                    %% remove subscribers...
                     Subscribers = mnesia:index_read(subscriber, DownPid, #mqtt_subscriber.pid),
                     lists:foreach(fun(Sub = #mqtt_subscriber{topic = Topic}) ->
                                 mnesia:delete_object(subscriber, Sub, write),
@@ -302,6 +371,9 @@ code_change(_OldVsn, State, _Extra) ->
 %%% Internal functions
 %%%=============================================================================
 
+add_queue(QueueR) ->
+    mnesia:write(queue, QueueR, write).
+
 add_topic(TopicR = #mqtt_topic{topic = Topic}) ->
     case mnesia:wread({topic, Topic}) of
         [] ->
@@ -314,9 +386,24 @@ add_topic(TopicR = #mqtt_topic{topic = Topic}) ->
             end
     end.
 
-add_subscriber({TopicR, Subscriber}) when is_record(TopicR, mqtt_topic) ->
+%% Fix issue #53 - Remove Overlapping Subscriptions
+add_subscriber({TopicR, Subscriber = #mqtt_subscriber{topic = Topic, qos = Qos, pid = SubPid}})
+                when is_record(TopicR, mqtt_topic) ->
     case add_topic(TopicR) of
         ok ->
+            OverlapSubs = [Sub || Sub = #mqtt_subscriber{topic = SubTopic, qos = SubQos}
+                                    <- mnesia:index_read(subscriber, SubPid, #mqtt_subscriber.pid),
+                                        SubTopic =:= Topic, SubQos =/= Qos],
+
+            %% remove overlapping subscribers
+            if
+                length(OverlapSubs) =:= 0 -> ok;
+                true ->
+                    lager:warning("Remove overlapping subscribers: ~p", [OverlapSubs]),
+                    [mnesia:delete_object(subscriber, OverlapSub, write) || OverlapSub <- OverlapSubs]
+            end,
+
+            %% insert subscriber
             mnesia:write(subscriber, Subscriber, write);
         Error -> 
             Error
@@ -330,6 +417,14 @@ monitor_subscriber(SubPid, State = #state{submap = SubMap}) ->
             SubMap
     end,
     State#state{submap = NewSubMap}.
+
+remove_queue(#mqtt_queue{name = Name, subpid = Pid}) ->
+    case mnesia:wread({queue, Name}) of
+        [R = #mqtt_queue{subpid = Pid}] -> 
+            mnesia:delete(queue, R, write);
+        _ ->
+            ok
+    end.
 
 remove_subscriber({TopicR, Subscriber}) when is_record(TopicR, mqtt_topic) ->
     [mnesia:delete_object(subscriber, Sub, write) || 
@@ -351,20 +446,37 @@ try_remove_topic(TopicR = #mqtt_topic{topic = Topic}) ->
 %%%=============================================================================
 %%% Stats functions
 %%%=============================================================================
+
 setstats(all) ->
-    setstats(topics),
-    setstats(subscribers);
+    [setstats(Stat) || Stat <- [queues, topics, subscribers]];
+
+setstats(queues) ->
+    emqttd_stats:setstats('queues/count', 'queues/max',
+                           mnesia:table_info(queue, size));
+
 setstats(topics) ->
-    emqttd_stats:setstat('topics/count',
+    emqttd_stats:setstats('topics/count', 'topics/max',
                           mnesia:table_info(topic, size));
 setstats(subscribers) ->
-    emqttd_stats:setstats('subscribers/count',
-                           'subscribers/max',
+    emqttd_stats:setstats('subscribers/count', 'subscribers/max',
                            mnesia:table_info(subscriber, size)).
 
+%%TODO: queue dropped?
 setstats(dropped, false) ->
     ignore;
 setstats(dropped, true) ->
     emqttd_metrics:inc('messages/dropped').
 
+
+%%%=============================================================================
+%%% Trace functions
+%%%=============================================================================
+
+trace(publish, From, _Msg) when is_atom(From) ->
+    %%dont' trace broker publish
+    ignore;
+
+trace(publish, From, #mqtt_message{topic = Topic, payload = Payload}) ->
+    lager:info([{client, From}, {topic, Topic}],
+                    "~s PUBLISH to ~s: ~p", [From, Topic, Payload]).