+++ apps/emqttd/src/emqttd_session.erl
@@ -173,17 +173,31 @@ puback(SessPid, {?PUBCOMP, PacketId}) when is_pid(SessPid) ->
 %%------------------------------------------------------------------------------
 -spec subscribe(session(), [{binary(), mqtt_qos()}]) -> {ok, session(), [mqtt_qos()]}.
 subscribe(SessState = #session_state{clientid = ClientId, submap = SubMap}, Topics) ->
-    Resubs = [Topic || {Name, _Qos} = Topic <- Topics, maps:is_key(Name, SubMap)], 
-    case Resubs of
-        [] -> ok;
-        _  -> lager:warning("~s resubscribe ~p", [ClientId, Resubs])
-    end,
-    SubMap1 = lists:foldl(fun({Name, Qos}, Acc) -> maps:put(Name, Qos, Acc) end, SubMap, Topics),
+
+    %% subscribe first and don't care if the subscriptions have been existed
     {ok, GrantedQos} = emqttd_pubsub:subscribe(Topics),
+
     lager:info([{client, ClientId}], "Client ~s subscribe ~p. Granted QoS: ~p",
-                    [ClientId, Topics, GrantedQos]),
-    %%TODO: should be gen_event and notification...
-    [emqttd_msg_store:redeliver(Name, self()) || {Name, _} <- Topics],
+                   [ClientId, Topics, GrantedQos]),
+
+
+    %% <MQTT V3.1.1>: 3.8.4
+    %% Where the Topic Filter is not identical to any existing Subscription’s filter,
+    %% a new Subscription is created and all matching retained messages are sent.
+    lists:foreach(fun({Name, _Qos}) ->
+          case maps:is_key(Name, SubMap) of
+          true ->
+              lager:warning("~s resubscribe ~p", [ClientId, Name]);
+          false ->
+              %%TODO: this is not right, rewrite later...
+              emqttd_msg_store:redeliver(Name, self())
+          end
+    end, Topics),
+
+    SubMap1 = lists:foldl(fun({Name, Qos}, Acc) ->
+                                  maps:put(Name, Qos, Acc)
+                          end, SubMap, Topics),
+
     {ok, SessState#session_state{submap = SubMap1}, GrantedQos};
 
 subscribe(SessPid, Topics) when is_pid(SessPid) ->
@@ -258,7 +272,7 @@ init([ClientId, ClientPid]) ->
     process_flag(trap_exit, true),
     %%TODO: Is this OK? or should monitor...
     true = link(ClientPid),
-    SessOpts = emqttd:env(session),
+    SessOpts = emqttd:env(mqtt, session),
     State = initial_state(ClientId, ClientPid),
     Expires = proplists:get_value(expires, SessOpts, 1) * 3600,
     MsgQueue = emqttd_queue:new(proplists:get_value(max_queue, SessOpts, 1000), 
@@ -360,7 +374,7 @@ handle_info({dispatch, {_From, Message}}, State) ->
 
 handle_info({'EXIT', ClientPid, Reason}, State = #session_state{clientid = ClientId,
                                                                 client_pid = ClientPid}) ->
-    lager:error("Session: client ~s@~p exited, caused by ~p", [ClientId, ClientPid, Reason]),
+    lager:info("Session: client ~s@~p exited for ~p", [ClientId, ClientPid, Reason]),
     {noreply, start_expire_timer(State#session_state{client_pid = undefined})};
 
 handle_info({'EXIT', ClientPid0, _Reason}, State = #session_state{client_pid = ClientPid}) ->