+++ CHANGELOG.md
@@ -2,6 +2,86 @@
 emqttd ChangeLog
 ==================
 
+0.8.6-beta (2015-06-17)
+-------------------------
+
+Bugfix: issue #175 - publish Will message when websocket is closed without 'DISCONNECT' packet
+
+
+0.8.5-beta (2015-06-10)
+-------------------------
+
+Bugfix: issue #53 - client will receive duplicate messages when overlapping subscription
+
+
+0.8.4-beta (2015-06-08)
+-------------------------
+
+Bugfix: issue #165 - duplicated message when publish 'retained' message to persistent client
+
+
+0.8.3-beta (2015-06-05)
+-------------------------
+
+Bugfix: issue #158 - should queue:in new message after old one dropped
+
+Bugfix: issue #155 - emqtt_parser.erl: parse_topics/3 should reverse topics
+
+Bugfix: issue #149 - Forget to merge plugins/emqttd_auth_mysql from 'dev' branch to 'master' in 0.8.x release
+
+
+0.8.2-alpha (2015-06-01)
+-------------------------
+
+Bugfix: issue #147 - WebSocket client cannot subscribe queue '$Q/queue/${clientId}'
+
+Bugfix: issue #146 - emqttd_auth_ldap: fill(Username, UserDn) is not right
+
+
+0.8.1-alpha (2015-05-28)
+-------------------------
+
+Client [Presence](https://github.com/emqtt/emqttd/wiki/Presence) Support and [$SYS Topics](https://github.com/emqtt/emqttd/wiki/$SYS-Topics) Redesigned!
+
+Bugfix: issue #138 - when client disconnected normally, broker will not publish disconnected $SYS message
+
+Bugfix: fix websocket url in emqttd/priv/www/websocket.html
+
+Improve: etc/emqttd.config to allow websocket connections from any hosts
+
+Improve: rel/reltool.config to exclude unnecessary apps.
+
+
+0.8.0-alpha (2015-05-25)
+-------------------------
+
+[Hooks](https://github.com/emqtt/emqttd/wiki/Hooks%20Design), Modules and [Plugins](https://github.com/emqtt/emqttd/wiki/Plugin%20Design) to extend the broker Now!
+
+Plugin: emqttd_auth_mysql - MySQL authentication plugin (issues #116, #120)
+
+Plugin: emqttd_auth_ldap - LDAP authentication plugin
+
+Feature: emqttd_broker to support Hooks API
+
+Feature: issue #111 - Support 'Forced Subscriptions' by emqttd_mod_autosub module
+
+Feature: issue #126 - Support 'Rewrite rules' by emqttd_mod_rewrite module
+
+Improve: Support hooks, modules to extend the broker
+
+Improve: issue #76 - dialyzer check
+
+Improve: 'Get Started', 'User Guide', 'Developer Guide' Wiki
+
+Improve: emqtt_topic to add join/1, feed_var/3, is_queue/1
+
+Improve: emqttd_pooler to execute common tasks
+
+Improve: add emqttd_sm_sup module, and use 'hash' gproc_pool to manage sessions
+
+Tests: add more test cases for 'emqttd' app
+
+
 0.7.1-alpha (2015-05-04)
 -------------------------